Makefile: assume MANDIR argument is already a manual directory
The MANDIR argument for secvarctl allows the user to set the location of the manual directory during installation. Previously, it was assuming the directory above the man directory, so it assumed it contained a subdirectory `man` and `man` contained another subdirectory `man1`. With this update, the MANDIR set by the user should point directly to a `man` directory. In other words if trying to install to `/usr/local/man/man1` , the user can now run `make install MANDIR=/usr/local/man` rather than `make install MANDIR=`/usr/local`. We feel this new implementation makes more sense from a user perspective