Add functionality for auth generation with signed data instead of private key
Previously, if a user wanted to make a signed auth file, they would need to supply secvarctl with public and private key pairs for signing the data. Recently, it has come to our attention that some users will not have access to their private keys as it may be under the guard of a signing framework server. In this case, the most secvarctl can do is to generate the hash to be signed. From their, it is up to the user to generate the signature with their signing framework. Then the user can give secvarctl the signature and the generation of PKCS7/auth files can resume. It is important the user uses the same timestamp in both the presignature generation and auth/PKCS7 generation. Now, with this commit, the user can use a signing framework to generate an auth file using the below commands:
    - $ secvarctl generate c:x -n <varName> -t <timeStamp> -i <certificate> -o <sha256.hash>
    - $ <user sends hash to be signed -in <sha256.hash> -out <signature.bin>
    - $ secvarctl generate c:a -n <varName> <timeStamp> -i <certificate> -o <newAuth> -c <signingCert> -s <signature.bin>