runSvcGenerateTests.py: Add tests for external signature
In the two previus commits, documentation and functionality were added to allow for secvarctl to accept an external signature rather than generating the signature internally. Now, we add tests to ensure that the enhanement was executed appropriately. To test we generate a signed auth file to work as our control. Then we generate an unsigned hash digest and sign it externally with openssl. We then use the signature to generate another signed auth file that will be our experiment. If the control and experiment contain the exact same data then the test passes.