Altered strcpy function calls
When building rpms, there are compilation flag warnings that are treated as errors. These errors all involve strncpy and strncat being called using a length that is based off of the source rather than the destination. To fix this, I changed the call to be strcpy (since it would be the same fucntion call if using strlen(source) for the length argument). After testing another rpm build, the warnings/errors are resolved. This commit does not impact the functionality of the code