test/Makefile: Remove generate data make trigger
Previously, if the generateTestData.py file were edited then the test data would be regenerated. This is not ideal because if even a comment is added to the generate python script, the entire testdata subdirectory would need to be recommitted. This is not ideal and the only reason why the data would need to be regenerated is if the user wanted it or if the user tried to run the tests and the test data did not exist. Now the test data can be generated if the user runs 'make generate'.