CMakeLists.txt: change default build type to Debug
This commit will change the default build type to debug. Bothe debug and coverage have been given additional flags: -Wall and -Werror. The reasoning is that most of the time that the project is built from source, it will be to incorporate/test changes or enhancements. This will encourage all warnings to be fixed before commiting a change. If building for general use, then the Release build type can be used (these flags have not been added here for this reason)