Start work on supporting openssl lib
This commit begins the process of abstracting the crypto functions into a higher layer. These abstract functions are stored in crypto.c and crypto.h . They allow for either openssl or mbedtls to be the backend crypto library to all the crypto functions. This work is necessary for secvarctl to be submitted to RHEL since mbedtls is not one of their accepted crypto libraries. Eventually, all of secvarctl will be able to be built with `make OPENSSL=1` . For now, this commit only adds functionality for the validate, read and write subcommands. Please note, `make OPENSSL=1` will not compile since the other functions have not been worked on yet. That being said building with Mbedtls (default) will still work. As openssl support increases in further commits, these files/functions will be ported until `make OPENSSL=1` compiles and executes correctly.