Port 'secvarctl verify' and skiboot commands to use abstract crypto functions
This commit has added the new openssl crypto calls to the secvarctl verify functions. Being that many of the secvarctl verify calls depend on the skiboot edk2-compat-process file, the skiboot files have also been edited to allow for the crypto functions. To reiterate, the new crypto functions allow for secvarctl to be compiled with either mbedtls or openssl. That being said, secvarctl will still not compile with `make OPENSSL=1` since the generate functions have not been ported yet. Compiling with mbedtls still works however. Ideally, the edits in the skiboot files will be reflected in upstream skiboot eventually. However, for now, any edits to the original skiboot code has been noted with something like "Nick Child edited because...".