Add openssl capability with secvarctl generate functions
This is commit allows for the generate function to also use the abstract crypto functions. As of this commit, all the secvarctl commands can now run with either mbedtls or openssl (make `OPENSSL=1`) as the crypto library. The only functionality that has not yet been incorporated is the generation of a pkcs7/auth using an externally generated signature 'secvarctl generate c:a -s ...' . I could not find any convienient openssl functions for adding a signature directly.