Manually add usage and help options to argp
Previously, in order to correctly exit from a --help or --usage flag, we had to do manually check for help and usage in the parse_opt function. It was difficult and messy since they are never actually sent directly to the function as `key` arguments. This led to a very unreadable and confusing block of code in all of the parse_opt functions. Thanks to the recomendations from Daniel Axtens, this commit manually checks for the usage and help flags, cleaning up the code and making things more readable.