Add reallocArray function
This commit adds a new function to the generic.c file. The new functions reallocArray, recieves a pointer to an array and reallocates it to have the requested amount of elements, where the size of each element has been specified. Essentially, reallocArray is just the realloc function with extra safeguards and better cleanup if an error occurs. It is currently only being called from parsing the `generate` subcommand. I felt it could be useful in the future so I added to the generic functions.

The reason for adding this function was because previously used `realloc` functions had issues with allocating and freeing memory. To avoid these potential bugs, the reallocArray function was implemented. This function ensures that the operation is successful and frees old memory if an error occurs. Thanks to Daniel Axtens for bringing this to our attention.