Change custom timestamp argument to ISO8601
This commit changes the previous custom timestamp option from '-t y-m-d h-m-s' to '-t y-m-dT-h-m-s' this is to align more with the ISO 8601 standard. Additionally, this commit changes some validation of the timestamp so that it better follows the requirements of the struct efi_time in edk2.h . Timestamp generation is now in UTC (instead of previously used localtime). Both the custom timestamp and current real time (if no -t flag) will be assumed to be in UTC. To reflect this change. documentation and tests have also been updated.

Signed-off-by: Nick Child <nick.child@ibm.com>