edk2-svc-verify.c: Add return code check on validateVarsArg
Previously, the return code for validate vars arg was not being checked and was assumed to always return successful. After looking at the function, it is possible for this function to fail and, while it is rare, it is possible for memory allocations to fail. This commit checks the return code and reacts appropriately.

Signed-off-by: Nick Child <nick.child@ibm.com>