Initial use of new clang-format file
This commit is the first use of `make format` which calls the `.clang-format` file on all of the secvarctl specific (not in `external`) source code. All source code should now meet the coding standard we have defined. Since most of the source files have been tweaked in various areas, any type of rebasing before this point will be difficult.

Signed-off-by: Nick Child <nick.child@ibm.com>