secvarctl: use get_esl_cert from skiboot
secvarctl had its own unique version that took a ESL instead of
just a buffer. That certainly has its merits, but for simplicity
and to avoid carrying 2 almost identical versions, drop the one
from secvarctl and use the one from skiboot.

We keep skiboot's assert(): it doesn't catch out-of-memory,
it catches calling the function with an invalid handle. (That is,
instead of passing a char ** pointer which we can dereference and
write the address of the new memory, it is NULL, so we can't write
the address of the new memory.)

We can afford to assert() here - it's not something we'd expect to
hit unless we made a serious error in coding; we don't expect any
amount of corrupt file data to be able to cause it.

Signed-off-by: Daniel Axtens <dja@axtens.net>