include/config.h: Use `make format` to reformat source code
This commmit is the result of running, `make format` on all of the non `external` secvarctl source code. Only the file `include/config.h` was affected. The purpose of running this command is to use the tool `clang-format` to ensure the source code is aligned with the codeing standard defined in the `.clang-format` file.

Signed-off-by: Nick Child <nick.child@ibm.com>