test: Update tests to pass with openssl builds
This commmit adds the flag `OPENSSL=1` to the Makefile in the `test` directory. This will only run the tests that are supposed to be run with Openssl. Previously, CI was failing since secvarctl built with openssl does not allow for external signatures for PKCS7 generation. This commmit allows us to omit testing that feature with `cd test && make OPENSSL=1`

Signed-off-by: Nick Child <nick.child@ibm.com>