skiboot/config.h: Remove config file since we have our own now
With the recent inclusion of a custom skiboot config file, located in `include/config.h` . We no longer have use for `external/skiboot/include/config.h`. This file seems to be lost and actually should be in `external/skiboot/ccan/` but after digging deeper, it seems to be originally used in skiboot only for running tests. It was likely useful to secvarctl prior to having its own `config.h`, that being said there is no use in having both and everything compiles and passes tests without it.

Signed-off-by: Nick Child <nick.child@ibm.com>