Add new command `insert`
This commit adds the initial implementation of the 6th secvarctl command,
`insert`. This general idea is that this command can add a new ESL to the
current ESL chain in a secure variable. Since the data is just appended
single-entry ESL's, all we really need to do is append the new data to the
current data and generate an auth file over the combined data buffers.

The command requires a new esl (with '-i'), a varname (with '-n') and auth
generation info for signing (at least one pair of signers with '-k'/'-x',
'-c'). It also requires either an '-w' or '-o' flag for either writing the
output data to the secvar sysfs `update` file or to a user defined output
file, respectively. Similarly, the command can read the current esl (the one
that is being appended to) from the secvar sysfs `data` file (default) or
from a user defined input file (with '-e'). The path to the sysfs is /sys/
firmware/secvar/vars/ by default but can be changed with '-p' option. A few
other flags like '-f', '-t', '-h' have also been added. Still to be added is
documentation and tests.

Signed-off-by: Nick Child <nick.child@ibm.com>