edk2-svc-insert.c: make argp child parser for future remove command
This commit factors out some argp parsing for future implementations of the `remove` command. Since, `remove` should essentially do the opposite as `insert`, it will take some of the same command line options. To avoid reusing code, we use `arg_child` structs again. This strategy was previously used with sharing flags between `secvarctl generate` and `secvarctl insert`.

Admitedly, this is extra confusing since the new shared parser for `insert` and `remove` has its own child parser for giving auth flags for signing. It makes use of the ability for a child parser to have a child parser. Think of these child parsers as the union of mulitple sets of command line options.

Signed-off-by: Nick Child <nick.child@ibm.com>