edk2-svc-insert.c: Expose required options in help message
Previously, commmand line options like '-o', '-i' and '-w' were hidden from the `secvactl insert --help` message. This was because they weren't exactly an option (since they were mandatory), so we marked them with the flag `OPTION_HIDDEN`. But this flag has been deemed unnessary because some description of the command line argument should be given and it should not be restrained in the help message. We do not want to treat them as normal options because then they will appear twice in `secvarctl insert --usage`, once in the list of options and again in the requirments. We will now use the flag OPTION_NO_USAGE, which will display the flag in the help message but not in the list of options when using `secvarctl insert --usage`.

Note: It would be nice to also label the required flags for auth signing {-n, -s/-k, -c} as OPTION_NO_USAGE. But since `secvarctl generate` argp parsers also using these flags through argp child parsers, we will leave them as regular options. This conclusion was because it is likely better to have those flags appear twice in `secvarctl insert --usage` as a usage and a requirement than to have them not appear in `secvarctl generate --usage` at all.

Signed-off-by: Nick Child <nick.child@ibm.com>