edk2-svc-insert: MAYBE SQUASH Don't allow insert or remove of dbx
This commit removes the ability to insert dbx entries. Currently the dbx is being read as multiple ESL's each w/ one entry. I have heard about the possibility of the dbx being one ESL with multiple entries. It has been awhile since I checked in on this. I need to get confirmation of what the dbx format is. It is very possible this commit will be squashed. For now, it is muche easier to implement these functions without having to worry about the dbx so this commit just throws an error if the variable veing edited is the dbx

Signed-off-by: Nick Child <nick.child@ibm.com>