esk2-svc-insert.c: Implement secvarctl remove command
This commit is the initial implementation of `secvarctl remove` the purpose of this command is to remove specific entries to an esl chain and generate
a new auth file with the remaining data. This is very similar to the `insert` command. The only difference in arguments is that `insert` accepts a new ESL file with `-i <file>` and `remove` accepts a serial number with `-x <sn>`. The serial number is from an X509 contained in the ESL chain.

Signed-off-by: Nick Child <nick.child@ibm.com>