add max_len feature to getDataFromFile
This commit adds a parameter to the `getDataFromFile` for specifying the max
number of bytes to read from a file. This commit is beneficial because it
allows for less repetitive data when reading a secvar since it can now call
this function instead of reimplementing it with a max size. Additionally,
bugs were coming up when reading /sys/firmware/secvar/format, since the file
can have a much larger size than actual data. Now we can specify how many
bytes to read instead of trying to read the whole file. Most callers of this
function will use SIZE_MAX for the max size argument if they just want to read
an entire file.

Signed-off-by: Nick Child <nick.child@ibm.com>
[dja: fixed some merge conflicts where we had torn out use of generic
      functions in `external/`]
Signed-off-by: Daniel Axtens <dja@axtens.net>