edk2-svc-validate.c: Allow empty timestamp in timestamp validation
When validating the timestamp variable (TS), we run through every entry and return and ensure that the values are logical. Having the year set to zero is invalid. But if the entire 16 byte timestamp is zero, then it technically is valid since it means that there is currently no entries for that secvar.
This commit adds a check to allow for empty ts entries when validating the TS variable. It also adds a test to ensure the bug is patched.

Signed-off-by: Nick Child <nick.child@ibm.com>