edk2-compat.c: Turn off setup mode if PK is updated
Previously, we are commenting out the section of code that enforced secure boot if the the PK is updated. I likely thought it was HW key related so I commented it out. I uncommented so now secure boot will be enforced as soon as a PK is enrolled.

Additionally, a test has been added to prove successful patching of the bug.

Signed-off-by: Nick Child <nick.child@ibm.com>