crypto-openssl.c: Unify return codes
Since `crypto_strerror` is called on the return code of
`crypto_pkcs7_signed_hash_verify`, we need to make sure that
an openssl error code is returned. Rather than change the API
for one function, we should really be always returning an openssl
error code. This commit changes all crypto functions to return an
openssl error code when secvarctl is compiled with openssl.

Since openssl never returns specific error codes, we make use of
their error stack with `ERR_get_error`. If the stack is empty,
then we make an artificial error code.

Signed-off-by: Nick Child <nick.child@ibm.com>