Update crypto calling functions to expect success return codes
The crypto abstraction functions have been updated to reuturn
error codes that are specific crypto library (gnutls/mbedtls/openssl).
The calling functions should expect to recieve a success value that is
backend specific.It just so happens that all of the current crypto
libraries return 0 on success so this commit doesn't cause any behavior
to change. But, it does add clarity to be more specific on the expected
return value. Also, in the future, if we decide to support another library
then we won't need to panic about its return value.

Signed-off-by: Nick Child <nick.child@ibm.com>