Add github action for CI
This introduces github actions to run CI tests on all new
pushes and pull requests to the repository. This workflow
will use `make` and `cmake` to build secvarctl with mbedtls
and openssl. It will run the python test scripts and run
memory checks with valgrind. These are done on Github's x86
Ubuntu runners.

Signed-off-by: Nick Child <nick.child@ibm.com>