Add github action for building rpm's
This commit will add a workflow that builds rpm's on every new
release of secvarctl. As a result, rpm's for x86 and ppc64le will
be generated, as well as a source rpm. These will then be uploaded
as an asset of the new release. These rpm's are built with openssl.

Since github actions does not support rpm based distro's in their runners,
we will add and run in a fedora docker container to build the rpm's.

Signed-off-by: Nick Child <nick.child@ibm.com>