runSvcGenerateTests.py: Add tests for remove command
This commit adds tests to the python script for the new `secvarctl remove
command`. The test works by making a KEK with many ESL entries. We then
generate a db auth file signed with an entry in the KEK and use `secvarctl
verify` to prove that an update would be successful. After using `secvarctl
remove` to remove the KEK entry, we see that the same db update would now
fail (since it's signer has been removed from the KEK). We continue through
this process until the KEK is empty.