Allow insert command on dbx and add tests
Previously, the dbx was not allowed to be used in a `secvarctl
insert` command. After reconsidering, since a dbx is still just
appended ESL's the insert opperation on a dbx is no different
than with the other secure variables.
To prove this point, tests were added to the test script.
NOTE: dbx `remove` opperations will not work

Signed-off-by: Nick Child <nick.child@ibm.com>