Allow remove to work with dbx and add tests
This commit allows for `secvarctl remove` to work properly
with the dbx. Since the dbx contains hashes instead of x509's,
we will use the hash values to identify the ESL that the user
would like to remove. For this reason, in the secvarctl command,
if the `-n <varName>` flag is the dbx, then the `-x XX:XX:...XX`
flag should contain the hash. If not dbx, then `-x` should still
be the serial number.

To prove the functionality, dbx test cases have been added to the
remove tests.

Signed-off-by: Nick Child <nick.child@ibm.com>