crypto: Fix openssl implementation of crypto_x509_is_RSA
The previous implementation was attempting to get the
type of signature used to sign the x509. It is now fixed,
to get the type of the public key. If the type of the key
cannot be determined then openssl will return NID_undef
which has a value of 0 (secvarctls SUCCESS value). If
this occurs then the implementation should simply return
a generic CERT_FAIL return value. The calling function
in edk2-svc-validate is adjusted to accept this generic return
value.

Signed-off-by: Nick Child <nick.child@ibm.com>