secvarctl: reorganised host secure boot variable backend
The host secure boot variable backend code was reorganised. The edk2-compat
code was moved to the host directory, and the -m option in secvarctl was
added to allow for the selection of the secure boot variable's mode.

By default, we generating host backend static library and linked it with
secvarctl binary. Also, provided the option to generating and linking of
host backend dynamic library

Ex: make DYNAMIC_LIB=1

Signed-off-by: Sudhakar Kuppusamy <sudhakar@linux.ibm.com>