guest/verify: rewrite get_current_esl()
In order to fix a pair of uninitialized variable warnings from
current_esl_size, the function has be reorganized to avoid a giant
if/else expression, and set consistent outputs for the pass-by-reference
arguments when exiting early.

Since current variables supplied with -c should take precedence over
variables from a path, the "else" condition for-loop has been moved to
the top of the function. If a matching variable is not located in the
current variable list, it falls through to checking for the variable in
the variable path.

Furthermore, all early returns have been replaced with goto out/fails
where appropriate, so that the pass-by-reference variables can be
returned consistently, which was the source of the uninitialized
variable warning.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>