Makefile: fix partial rebuilds when header files change
A previous rework commit either removed the inclusion of dependency
files, or never existed in the first place. Re-add that so changes to
headers can properly trigger a rebuild.

Also remove the useless `coverage` build target, which current is just
a phony alias for `bin/secvarctl-cov`. This will return later when there
is actual coverage testing implemented.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>