guest/read: reduce scope of loop-only variables
cppcheck flags variables that can be reduced in scope, meaning they are
declared in say the top-level of a function, but only used in the scope
of a loop.

This in itself isn't an issue, as I personally prefer seeing all the
variables used in the function as a whole up at the top, however it does
highlight a larger issue with the function as a whole -- FAR too much
indentation and probably repeat logic.

This function should *really* be broken down into helper functions so
that it need not be one massive while loop.

For now, reduce the variables to the scopes that they are used in.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>