guest/verify: functions should not just be one giant conditional
get_pk_and_kek_from_path_var previously contained one giant `if` block
with all of the logic.

Invert the logic and return early, so that the rest of the body can
have the indentation reduced. Also return an error code instead of
SUCCESS if that condition is not met, since apparently it just happily
returns on through without doing anything.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>