guest/read: size_t variables literally can't be -1
Fix check of esl_data_size against -1, which will never happen. Since
this appears to be checking that there is enough space in the esl buffer
for the timestamp header + data, instead just check that the size is at
least the length of a timestamp. If the size is less than a timestamp,
it's probably garbage data.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>