host/read: fix weird pointer increment in readTS
For some reason the for loop uses an assigment as its condition. I'm
going to ignore that and address the portability issue of casting a
pointer to void then doing arithmetic on a non-known size for `void *`.
Just increment the pointer if you are only going to add the size of the
value anyway.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>