guest/verify: remove extraneous code after break in parse_options
I'm going to be honest, I have no idea where that code came from, or
why it was there in the first place. It appears the host equivalent to
this function also has code after the break, but cppcheck isn't flagging
that for some reason.

Adding a default: to the switch breaks a test. Removing the initial
break breaks a test. None of this make sense, so just remove the
offending code, and improve the test cases later -- maybe we'll find out
why this is so miserable later.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>