host/generate: check that ESL is not NULL is getPreHashForSecVar
Reported by scan-build.

ESL is not checked for NULL throughout the function, and is blatantly
memcpy'd from at the end from an also unchecked size value. This is
however, actually valid, the case of ESL_size = 0, as the memcpy will
then not actually read from the buffer. Even better, this is exactly how
generating a reset auth file works -- it passes in a NULL buffer with
ESL_size set to 0.

Therefore, this commit adds a check for that specific error case where
ESL is NULL, but ESL_size is nonzero. Also this skips the memcpy at the
end in the case of ESL being NULL, just to ensure that any static
analyzers are happy about not blatantly memcpy'ing from NULL, even if
ESL_size = 0 would prevent it.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>