guest/verify: fix potential double free in get_pk_and_kek from update_var
Reported by scan-build.

Currently, this attempts to free auth_data as returned by
get_auth_data() if rc is non-success. get_auth_data() only sets
*auth_data if returning successfully. Therefore, auth_data will never be
assigned to an allocated buffer in this particular conditional block.

Remove the free, as it is not necessary. Functions that return allocated
memory should really be marked better.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>