CI: use apt-get instead of apt, update before install
Potential quirk of using act to run workflows locally, but there should
be no assumption that the repository has been updated before
installation. It may not be a bad idea to define a custom containerfile
for this later.

Also use apt-get instead of apt, as apt does not have a stable
scriptable API, and likes to complain about it.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>