Makefile: rework debug build, replace valgrind with ASAN
Currently, there are two binaries that the Makefile can build:
secvarctl, and secvarctl-cov. The latter was intended for coverage
testing, and therefore is the binary used by the test cases.

However, in the effort to integrate address/ub santizers, this would
have created a third build type: release, coverage, and memcheck.

Instead, change the coverage build to be a generic "debug" build
binary, that integrates coverage testing, address santizing, debug
symbols, etc. This also replaces the need for a separate "memcheck"
testing target, as memcheck will now always occur when running the
tests.

Also updates the default path in the test cases, that will probably be
factored out in a later patch.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>