Makefile: add variable-based control for disabling ASAN
ASAN is great for finding problems, but a bit of a pain when trying to
fix one problem while there are still more to address, particularly if
attempting to avoid regressions.

As a secondary issue, it depends on the availability of libasan and
libubsan, which may not typically be installed by default.

This patch adds the DISABLE_ASAN option, which may be set to 1 to skip
building with ASAN support.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>