guest/util: rewrite get_signature_type() to use a comprehensive lookup table
ASAN really doesn't like the amount of poorly bounds-checked memcmps in
this file. In a first attempt to mitigate the usage of memcmp for
comparing strings, first collect all the possible signature types into
one array so that get_signature_type() can iterate over all of them in
a single loop.

This introduces an enumeration with a unique name for each type, that
may be used in the future instead of a string, to entirely remove the
need for the string comparisons used elsewhere in the code.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>