Makefile: change check target dependency to depend on the binary to be tested
Previously, it would explicitly depend on the pseudotarget
secvarctl-dbg, so that running `make check` would ensure that the debug
binary was up-to-date on each invocation. However, this is a bit
inflexible for running the tests against another binary (particularly
in the case of a CMake build).

Instead, have the check target depend on the binary pointed to by
SECVAR_TOOL, and set the default value to the original debug binary.
This allows for the regular default make check behavior to remain the
same, but also has the added benefit of testing another binary path
while also implicitly checking that the binary exists due to dependency
requirements.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>