test/common: factor out common functions into a shared base test class
This commit starts the process of factoring out some shared logic
between the test cases for guest and host. This is an intermediate
commit -- future commits WILL remove or rework code that may have been
altered here. This is to avoid all the changes snowballing into one
massive rework commit.

That said, unfortunately a directory rename got wrapped up in this
commit -- rather than implement a parameter that would probably be
removed later, just rename the directory so that both match. Expect
this to change again later.

Anyway, the tl;dr is that `common.py` now defines a `SecvarctlTest`
class for the unit tests to derive from, and for now factors out the
command running and environment set-up logic. There will be more.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>