test/common: add generic assertCmd() for taking Success/Fail as an argument
The host tests use a large table of commands and expected results, so
rather than use some dumb reflection to call the right function, just
create a function for those.

Factor out the logic in assertCmdTrue and assertCmdFalse to this
function, so those are now just simple wrappers to a single assert
function.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>