test/common: remove all command output logging
Each test file was very inconsistent about how it handled output
logging, and also cause a lot of repeated code. Furthermore, it made
debugging a failing test somewhat difficult, since the developer would
have to figure out which log file would have the relevant output. This
is also a significant concern for CI, since the output of the job will
not display the relevant failure.

This commit replaces all use of the logging with a new system that
prints the output of a command if the related test case fails. This
extends the current assertCmd family of functions to inject the output
into the assert message.

NOTE: this commit also removes the PPC-specific test case runners since
that also had some custom logging behavior. If we want to keep
PPC-specific test cases, they should be in a separate test file not
included in the default runners.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>