Makefile: reenable single-backend builds
Building with support for only a single backend was lost in the massive
overhaul of the build system. Re-enable the ability to build for only
guest or host.

This slightly rearranges the Makefile a bit -- backend include Makefiles
should now contain all their dependencies, subsequent commits will
make further progress on fully decoupling the two backends.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>