Makefile: prepare for building a different flavor of libstb-secvar based on crypto library
Right now, the build depends on a hardcoded version of libstb-secvar
as "libstb-secvar-openssl.a". Split up the current variable into chunks
that can be individually tweaked for the version of libstb-secvar that
should be built.

A bonus side-effect is that LIBSTB_SECVAR_ROOT can now be overridden to
use a local clone instead of the submodule.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>