guest/common: Replace hash_funct_t with index to list
Not sure how I feel about this but since hash_func_t has been removed
from libstb-secvar, and the equivalent information is now being
stored in signature_type_list. Have the guest functions pass around
an enum that works as an index to signature type list. This will
act a a token to access certain information about the algorithm
later on. It did not come off as clean as I would have liked.

Also as of this commit, don't expect things to compile. Still
chuggin through changes.

Signed-off-by: Nick Child <nick.child@ibm.com>