secvarctl: rewrite backend selection logic
There were a few goals with this rewrite:
1. be able to detect which backends are enabled at runtime
2. allow selecting backend by shortname "host" or format "ibm,edk-compat-v1"
3. be consistent in how the backends are looked up, managed, and added

The command table is a good way to get quick access to a specific
backend. This commit improves the lookup by tying the index of each
entry in the table to an enum, so that enabling/disabling backends keeps
the index in the table a consistent way to get a specific backend's
information without having to tie it down with a variable name (and
thus, more ifdefs). In a similar related change, a table tying
strings -> backend enum has been added to simplify lookup of a backend
given a string, while also allowing for backend aliases.

All of this now means that (in theory) it should be easier to add/remove
backends in the future by only adding the relevant information into
the table, and the rest is factored out in the main body of code.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>