test: only run tests if secvarctl with built with the correct backend
When building with only a single backend, `make check` will no longer
work since it will still attempt to run e.g. host tests on a guest-only
build.

This commit adds a common function that checks if secvarctl is built
with the expected backend, and skips the tests if not.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>