CI: split out building/testing and static analysis into reusable workflows
This commit splits out the push/pr workflows into reusable workflows
so that these processes can be integrated into future workflows without
the need for duplicated effort. The main intent is for the
release-oriented workflow to be able to also depend on a successful run
of the tests before even attempting to run rpmbuild.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>