secvarctl: display enabled backends in --help, allow empty -m to query enabled backends
Currently, there is no way to know which backends are enabled in a given
binary. Therefore, add the enabled backends to the --help output, as
well as display the enabled backends if -m is called without any
following argument.

This also fixes a bug where argv was checked for NULL instead of *argv,
which could have led to a segmentation fault.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>