spec: slight build tweaks for maximum distro compatibility
Some of the rpmbuild macros enable FORTIFY_SOURCE, which requires
building with optimizations. Currently, the specfile does not define
a build type, therefore it uses the unoptimized Debug profile. Change
the %cmake invocation to include the build type.

It also appears that SUSE uses a different method of building in
%cmake_build that caused it to build using the Makefile build instead
of the generated cmake build files, only to then build it again during
the %install stage. Move the %cmake call into the %build stage with
the %cmake_build to appease the cmake/rpmbuild gods.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>