guest/read: allow paths with or without a trailing slash
Fixes #62.

Paths supplied to read are expected to contain a trailing slash, which
is a regression from previous versions. Change get_variable_path to
automatically insert a trailing slash if needed.

Note: this function is a candidate for factoring out into common code,
I suspect this behavior might be present in other bits of backend code.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>