guest/generate: change --append option to be a boolean based on presence
Fixes #65.

Currently, the -a --append option accepts a string to be parsed into
an integer, either 0 or 1. This integer is then used to set a boolean
append flag.

Avoid parsing strings at all by having the boolean append flag be set
to true if -a/--append exists, otherwise default to false.

NOTE: This does introduce a change in the command line API.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>