guest/read: rewrite print_variables function to use libstb-secvar helper functions
Fixes #63, and maybe #61.

As reported in #63, a fuzzed ESL file causes a segfault when reading.
This occurs because the fuzzed ESL contains an internal size value that
is far larger than that of the ESL file itself. Therefore, when we hand
the data to OpenSSL to parse, we give the parsing function a very
incorrect size value to expect, and therefore it overruns the buffer.

Rather than add in more size checks, the function has been rewritten to
use the ESL/ESD iteration helper functions in libstb-secvar, which
already have coverage testing.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>